+++ presctl/cmd/delete.go
@@ -17,12 +17,13 @@ var deleteCmd = &cobra.Command{
 
 var delTemplateCmd = &cobra.Command{
 	Use:   "template",
+	Args:  cobra.MinimumNArgs(1),
 	Short: "delete a template resource",
 	Long:  `Use this command to delete a template from presidio.`,
 	Run: func(cmd *cobra.Command, args []string) {
 		actionName := getFlagValue(cmd, actionFlag)
 		projectName := getFlagValue(cmd, projectFlag)
-		templateName := getFlagValue(cmd, templateFlag)
+		templateName := args[0]
 
 		// Send a REST command to presidio instance to delete the requested template
 		entities.DeleteTemplate(&http.Client{}, projectName, actionName, templateName)
@@ -32,10 +33,11 @@ var delTemplateCmd = &cobra.Command{
 // delRecognizerCmd represents the recognizer command
 var delRecognizerCmd = &cobra.Command{
 	Use:   "recognizer",
+	Args:  cobra.MinimumNArgs(1),
 	Short: "delete an existing recognizer resource",
 	Long:  `Use this command to delete a recognizer.`,
 	Run: func(cmd *cobra.Command, args []string) {
-		recognizerName := getFlagValue(cmd, recognizerFlag)
+		recognizerName := args[0]
 
 		// Send a REST command to presidio instance to get the requested template
 		entities.DeleteRecognizer(&http.Client{}, recognizerName)
@@ -47,20 +49,12 @@ func init() {
 	deleteCmd.AddCommand(delTemplateCmd)
 
 	// define supported flags for the del command
-	delTemplateCmd.Flags().String(templateFlag, "", "template's name")
 	delTemplateCmd.Flags().String(actionFlag, "", "the requested action. Supported actions: ["+getSupportedActions()+"]")
 	delTemplateCmd.Flags().String(projectFlag, "", "project's name")
 
 	// mark flags as required
-	delTemplateCmd.MarkFlagRequired(templateFlag)
 	delTemplateCmd.MarkFlagRequired(actionFlag)
 	delTemplateCmd.MarkFlagRequired(projectFlag)
 
 	deleteCmd.AddCommand(delRecognizerCmd)
-
-	// define supported flags for the del command
-	delRecognizerCmd.Flags().String(recognizerFlag, "", "recognizer's name")
-
-	// mark flags as required
-	delRecognizerCmd.MarkFlagRequired(recognizerFlag)
 }
