+++ presctl/cmd/get.go
@@ -18,13 +18,14 @@ var getCmd = &cobra.Command{
 // templateCmd represents the template command
 var getTemplateCmd = &cobra.Command{
 	Use:   "template",
+	Args:  cobra.MinimumNArgs(1),
 	Short: "gets an existing template resource",
 	Long:  `Use this command to get a template.`,
 	Run: func(cmd *cobra.Command, args []string) {
 		actionName := getFlagValue(cmd, actionFlag)
 		outputFile := getFlagValue(cmd, outputFlag)
 		projectName := getFlagValue(cmd, projectFlag)
-		templateName := getFlagValue(cmd, templateFlag)
+		templateName := args[0]
 
 		// Send a REST command to presidio instance to get the requested template
 		entities.GetTemplate(&http.Client{}, projectName, actionName, templateName, outputFile)
@@ -34,11 +35,12 @@ var getTemplateCmd = &cobra.Command{
 // getRecognizerCmd represents the recognizer command
 var getRecognizerCmd = &cobra.Command{
 	Use:   "recognizer",
+	Args:  cobra.MinimumNArgs(1),
 	Short: "gets an existing recognizer resource",
 	Long:  `Use this command to get a recognizer.`,
 	Run: func(cmd *cobra.Command, args []string) {
 		outputFile := getFlagValue(cmd, outputFlag)
-		recognizerName := getFlagValue(cmd, recognizerFlag)
+		recognizerName := args[0]
 
 		// Send a REST command to presidio instance to get the requested template
 		entities.GetRecognizer(&http.Client{}, recognizerName, outputFile)
@@ -50,22 +52,16 @@ func init() {
 	getCmd.AddCommand(getTemplateCmd)
 
 	// define supported flags for the get command
-	getTemplateCmd.Flags().String(templateFlag, "", "template's name")
 	getTemplateCmd.Flags().StringP(outputFlag, "o", "", "path to the output file")
 	getTemplateCmd.Flags().String(actionFlag, "", "the requested action. Supported actions: ["+getSupportedActions()+"]")
 	getTemplateCmd.Flags().String(projectFlag, "", "project's name")
 
 	// mark flags as required
-	getTemplateCmd.MarkFlagRequired(templateFlag)
 	getTemplateCmd.MarkFlagRequired(actionFlag)
 	getTemplateCmd.MarkFlagRequired(projectFlag)
 
 	getCmd.AddCommand(getRecognizerCmd)
 
 	// define supported flags for the get command
-	getRecognizerCmd.Flags().String(recognizerFlag, "", "recognizer's name")
 	getRecognizerCmd.Flags().StringP(outputFlag, "o", "", "path to the output file")
-
-	// mark flags as required
-	getRecognizerCmd.MarkFlagRequired(recognizerFlag)
 }
